doc: add llms output for changelog & links
