+++ packages/docs/content/docs/meta.json
@@ -20,6 +20,7 @@
     "seo",
     "tips-tricks",
     "about",
+    "[Changelog](/docs/changelog)",
     "migrations/v2"
   ]
 }
